# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "VolleyPass API",
        "description": "API para el sistema de gestión de voleibol VolleyPass",
        "contact": {
            "name": "VolleyPass Support",
            "email": "admin@volleypass.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://my-default-host.com",
            "description": "VolleyPass API Server"
        }
    ],
    "paths": {
...
        "/api/public/matches/{id}/players": {
            "get": {
                "tags": [
                    "Public Match Players"
                ],
                "summary": "Obtener listado de jugadores de un partido",
                "operationId": "de3ebede9925544b88c36df97cceaf29",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del partido",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Listado de jugadores del partido",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "properties": {
                                                "match_id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "match_date": {
                                                    "type": "string",
                                                    "format": "date-time",
                                                    "example": "2024-01-15T15:30:00Z"
                                                },
                                                "status": {
                                                    "type": "string",
                                                    "example": "scheduled"
                                                },
                                                "tournament_name": {
                                                    "type": "string",
                                                    "example": "Copa Nacional 2024"
                                                },
                                                "venue_name": {
                                                    "type": "string",
                                                    "example": "Polideportivo Central"
                                                },
                                                "home_team": {
                                                    "properties": {
                                                        "team_id": {
                                                            "type": "integer",
                                                            "example": 5
                                                        },
                                                        "team_name": {
                                                            "type": "string",
                                                            "example": "Equipo A"
                                                        },
                                                        "logo": {
                                                            "type": "string",
                                                            "example": "https://example.com/logo.png"
                                                        },
                                                        "players": {
                                                            "type": "array",
                                                            "items": {
                                                                "properties": {
                                                                    "player_id": {
                                                                        "type": "integer",
                                                                        "example": 10
                                                                    },
                                                                    "name": {
                                                                        "type": "string",
                                                                        "example": "María García"
                                                                    },
                                                                    "jersey_number": {
                                                                        "type": "integer",
                                                                        "example": 7
                                                                    },
                                                                    "position": {
                                                                        "type": "string",
                                                                        "example": "Libero"
                                                                    },
                                                                    "is_captain": {
                                                                        "type": "boolean",
                                                                        "example": false
                                                                    },
                                                                    "is_starter": {
                                                                        "type": "boolean",
                                                                        "example": true
                                                                    },
                                                                    "birth_date": {
                                                                        "type": "string",
                                                                        "format": "date",
                                                                        "example": "1995-03-15"
                                                                    },
                                                                    "height": {
                                                                        "type": "integer",
                                                                        "example": 175
                                                                    },
                                                                    "weight": {
                                                                        "type": "integer",
                                                                        "example": 65
                                                                    }
                                                                },
                                                                "type": "object"
                                                            }
                                                        }
                                                    },
                                                    "type": "object"
                                                },
                                                "away_team": {
                                                    "properties": {
                                                        "team_id": {
                                                            "type": "integer",
                                                            "example": 8
                                                        },
                                                        "team_name": {
                                                            "type": "string",
                                                            "example": "Equipo B"
                                                        },
                                                        "logo": {
                                                            "type": "string",
                                                            "example": "https://example.com/logo2.png"
                                                        },
                                                        "players": {
                                                            "type": "array",
                                                            "items": {
                                                                "properties": {
                                                                    "player_id": {
                                                                        "type": "integer",
                                                                        "example": 15
                                                                    },
                                                                    "name": {
                                                                        "type": "string",
                                                                        "example": "Ana López"
                                                                    },
                                                                    "jersey_number": {
                                                                        "type": "integer",
                                                                        "example": 12
                                                                    },
                                                                    "position": {
                                                                        "type": "string",
                                                                        "example": "Atacante"
                                                                    },
                                                                    "is_captain": {
                                                                        "type": "boolean",
                                                                        "example": true
                                                                    },
                                                                    "is_starter": {
                                                                        "type": "boolean",
                                                                        "example": true
                                                                    },
                                                                    "birth_date": {
                                                                        "type": "string",
                                                                        "format": "date",
                                                                        "example": "1993-07-22"
                                                                    },
                                                                    "height": {
                                                                        "type": "integer",
                                                                        "example": 180
                                                                    },
                                                                    "weight": {
                                                                        "type": "integer",
                                                                        "example": 70
                                                                    }
                                                                },
                                                                "type": "object"
                                                            }
                                                        }
                                                    },
                                                    "type": "object"
                                                },
                                                "match_officials": {
                                                    "properties": {
                                                        "referee": {
                                                            "type": "string",
                                                            "example": "Carlos Mendoza"
                                                        },
                                                        "assistant_referee": {
                                                            "type": "string",
                                                            "example": "Luis Rodríguez"
                                                        },
                                                        "scorer": {
                                                            "type": "string",
                                                            "example": "Elena Vargas"
                                                        }
                                                    },
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Partido no encontrado"
                    }
                }
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Usar el token de autenticación en el header Authorization: Bearer {token}",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}